🚚 [Move]: PdfVersion を pdf/types/ から pdf/version/ に移動
PdfVersion は SUPPORTED_PDF_VERSIONS allowlist や compare など PDF 固有の
ドメイン知識を抱えており、refined primitive のみを集めた pdf/types/ よりも、
pdf/type/ や pdf/filter/ と並ぶドメイン概念ディレクトリ pdf/version/ が
適切。ByteOffset/ObjectNumber/GenerationNumber は引き続き types/ 残留。

- pdf/types/pdf-version/ → pdf/version/ に移動
- 既存インポートパスを更新 (catalog-parser, test.helpers)
- pdf/types/index.ts から再エクスポートを削除
- pdf/index.ts に version バレルを追加

Refs #16
