+++ packages/core/src/document/catalog-parser.ts
@@ -10,7 +10,7 @@ import type {
   PdfValue,
   TrailerDict,
 } from "../pdf/types/pdf-types/index";
-import { PdfVersion } from "../pdf/types/pdf-version/index";
+import { PdfVersion } from "../pdf/version/index";
 import type { Result } from "../utils/result/index";
 import { err, ok } from "../utils/result/index";
