+++ packages/core/src/pdf/index.ts
@@ -2,3 +2,4 @@ export * from "./errors/index";
 export * from "./filter/index";
 export * from "./type/index";
 export * from "./types/index";
+export { PdfVersion } from "./version/index";
