+++ packages/core/src/pdf/types/index.ts
@@ -2,4 +2,3 @@ export { ByteOffset } from "./byte-offset/index";
 export { GenerationNumber } from "./generation-number/index";
 export { ObjectNumber } from "./object-number/index";
 export * from "./pdf-types/index";
-export { PdfVersion } from "./pdf-version/index";
