+++ packages/core/src/pdf/version/index.ts
@@ -1,6 +1,6 @@
-import type { Brand } from "../../../utils/brand/index";
-import type { Result } from "../../../utils/result/index";
-import { err, ok } from "../../../utils/result/index";
+import type { Brand } from "../../utils/brand/index";
+import type { Result } from "../../utils/result/index";
+import { err, ok } from "../../utils/result/index";
 
 declare const PdfVersionBrand: unique symbol;
